core: Add FileReaderIterator
