+++ modules/core/src/iterator-utils/async-iteration.js
@@ -1,4 +1,5 @@
 /* global TextDecoder, TextEncoder */
+import {concatenateArrayBuffers} from '../javascript-utils/memory-copy-utils';
 
 // GENERAL UTILITIES
 
@@ -29,66 +30,23 @@ export async function forEach(iterator, visitor) {
   }
 }
 
-// ITERATOR GENERATORS
-
-// TextDecoder iterators
-// TextDecoder will keep any partial undecoded bytes between calls to `decode`
-
-export async function* textDecoderAsyncIterator(arrayBufferIterator, options) {
-  const textDecoder = new TextDecoder(options);
-  for await (const arrayBuffer of arrayBufferIterator) {
-    yield typeof arrayBuffer === 'string'
-      ? arrayBuffer
-      : textDecoder.decode(arrayBuffer, {stream: true});
-  }
-}
-
-// TextEncoder iterator
-// TODO - this is not useful unless min chunk size is given
-// TextEncoder will keep any partial undecoded bytes between calls to `encode`
-// If iterator does not yield strings, assume arrayBuffer and return unencoded
-
-export async function* textEncoderAsyncIterator(textIterator, options) {
-  const textEncoder = new TextEncoder();
-  for await (const text of textIterator) {
-    yield typeof text === 'string' ? textEncoder.encode(text) : text;
-  }
-}
+// Breaking big data into iterable chunks, concatenating iterable chunks into big data objects
 
 /**
- * @param textIterator async iterable yielding strings
- * @returns an async iterable over lines
- * See http://2ality.com/2018/04/async-iter-nodejs.html
+ * Concatenates all data chunks yielded by an (async) iterator
+ * Supports strings and ArrayBuffers
+ *
+ * This function can e.g. be used to enable atomic parsers to work on (async) iterator inputs
  */
-
-export async function* lineAsyncIterator(textIterator) {
-  let previous = '';
-  for await (const textChunk of textIterator) {
-    previous += textChunk;
-    let eolIndex;
-    while ((eolIndex = previous.indexOf('\n')) >= 0) {
-      // line includes the EOL
-      const line = previous.slice(0, eolIndex + 1);
-      previous = previous.slice(eolIndex + 1);
-      yield line;
+export async function concatenateChunksAsync(asyncIterator) {
+  let arrayBuffer = new ArrayBuffer(0);
+  let string = '';
+  for await (const chunk of asyncIterator) {
+    if (typeof chunk === 'string') {
+      string += chunk;
+    } else {
+      arrayBuffer = concatenateArrayBuffers(arrayBuffer, chunk);
     }
   }
-
-  if (previous.length > 0) {
-    yield previous;
-  }
-}
-
-/**
- * @param lineIterator async iterable yielding lines
- * @returns async iterable yielding numbered lines
- *
- * See http://2ality.com/2018/04/async-iter-nodejs.html
- */
-export async function* numberedLineAsyncIterator(lineIterator) {
-  let counter = 1;
-  for await (const line of lineIterator) {
-    yield {counter, line};
-    counter++;
-  }
+  return string || arrayBuffer;
 }
