+++ modules/core/test/iterator-utils/stream-iterator.spec.js
