+++ modules/ply/src/index.js
@@ -1,5 +1,2 @@
 // Import the worker bundled by webpack
 export {PLYLoader, PLYWorkerLoader} from './ply-loader';
-
-// TODO - PLYStreamLoader should eventually merge with/replace PLYLoader
-export {default as _PLYStreamLoader} from './ply-stream-loader';
